+++ package.json
@@ -10,6 +10,7 @@
   "scripts": {
     "test": "jest --config jestconfig.json",
     "build": "tsc",
+    "start": "tsc --watch",
     "clean": "rm -rf lib",
     "format": "prettier --write \"src/**/*.ts\" \"src/**/*.js\"",
     "lint": "eslint --fix \"src/**/*.ts\"",
